API

Modules

FunctionChains.FunctionChains

Types and constants

FunctionChains.FunctionChain

Functions and macros

FunctionChains.fchain

FunctionChains.with_intermediate_results

Documentation

FunctionChains.FunctionChains — Module

module FunctionChains

Implements chained functions (composed functions) beyond Base.ComposedFunction.

FunctionChains.FunctionChain — Type

struct FunctionChain{FS}<:Function

Represents a chain of composed functions.

A FunctionChain has a single field fs which may be a tuple, array or generator/iterator of functions.

(chain::FunctionChain)(x) applies the functions in the cain in order of iteration over chain.fs. ```

Supports with_intermediate_results. Also supports InverseFunctions.inverse and/or ChangesOfVariables.with_logabsdet_jacobian if all functions in the chain do so.

Use fchain to construct function chains instead of FunctionChain(fs).

FunctionChains.fchain — Function

fchain()
fchain(fs)

Construct a function chain of functions fs.

Typically returns a FunctionChain, but may be specialized to return other function chain types for specific types of functions.

fs must be iterable, it may be a tuple, vector, generator, etc. fchain(fs)(x) will apply the functions in fs in order of iteration.

FunctionChains.with_intermediate_results — Function

with_intermediate_results(f, x)

Apply multi-step function f to x and return a collection that contains the intermediate results and the final result.